Mark Token pointer as non-escaping via noescape()
The Go compile would move the Token to the heap
because of the EncoderMiddleware.EncodeToken()
interface call, which it cannot prove to store
the Token somewhere else in memory (in which case
it would escape).
We just establish a contract for any types implementing
that interface to _not_ store the Token.
This contract also exists with other information
store in the Token fields, which will also go
out of lifetime when the encoded XML element goes
out of scope in the encoder.

see: https://go.googlesource.com/go/+/go1.17.6/src/runtime/stubs.go#164