Improve decoder/encoder perf using pointer to Token
For decoding, providing a pointer to the NextToken()
method now eliminates all runtime.duffcopy() calls
which moved Token values around on the stack.
Now, the client provides a pointer to a Token value
which will be written to.

For encoding, providing a pointer does not necessarily
convey "this method mutates the token", but it is
also useful here to avoid copying at runtime.